backend/requirements.txt
Django==4.2.18
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg[binary]==3.2.6
